refactor: ViewHolder 클래스 삭제 및 내부에서 생성
