+++ app/src/main/java/woowacourse/movie/movielist/MovieListViewAdapter.kt
@@ -6,9 +6,12 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.BaseAdapter
-import woowacourse.movie.moviedetail.MovieDetailActivity
+import android.widget.Button
+import android.widget.ImageView
+import android.widget.TextView
 import woowacourse.movie.R
 import woowacourse.movie.domain.movieinfo.Movie
+import woowacourse.movie.moviedetail.MovieDetailActivity
 import java.time.format.DateTimeFormatter
 
 class MovieListViewAdapter(private val context: Context, private val movies: List<Movie>) :
@@ -61,6 +64,15 @@ class MovieListViewAdapter(private val context: Context, private val movies: Lis
             intent.putExtra(MOVIE_KEY, item)
             context.startActivity(intent)
         }
+    }
+
+    private class MovieListViewHolder(itemView: View) {
+        val moviePoster: ImageView = itemView.findViewById(R.id.movie_poster)
+        val movieTitle: TextView = itemView.findViewById(R.id.movie_title)
+        val screeningStartDate: TextView = itemView.findViewById(R.id.screening_start_date)
+        val screeningEndDate: TextView = itemView.findViewById(R.id.screening_end_date)
+        val runningTime: TextView = itemView.findViewById(R.id.running_time)
+        val bookButton: Button = itemView.findViewById(R.id.book_button)
     }
 
     companion object {
